+++ libraries/resource_mac_app_store_app.rb
@@ -42,6 +42,12 @@ class Chef
       #
       property :system_user, String, default: Etc.getlogin, desired_state: false
 
+      #
+      # If circumstances require, the reattach-to-user-namespace utility can be
+      # used every time we shell out to Mas.
+      #
+      property :use_rtun, [TrueClass, FalseClass], default: false
+
       ######################################################################
       # Every property below this point is for tracking resource state and #
       # should *not* be overridden.                                        #
@@ -72,8 +78,15 @@ class Chef
         converge_if_changed :installed do
           app_id = MacAppStore::Helpers::App.app_id_for?(new_resource.app_name)
           raise(Exceptions::InvalidAppName, new_resource.app_name) unless app_id
+
+          cmd = if new_resource.use_rtun
+                  include_recipe 'reattach-to-user-namespace'
+                  "reattach-to-user-namespace mas install #{app_id}"
+                else
+                  "mas install #{app_id}"
+                end
           execute "Install #{new_resource.app_name} with Mas" do
-            command "mas install #{app_id}"
+            command cmd
             user new_resource.system_user
           end
         end
@@ -87,8 +100,15 @@ class Chef
         converge_if_changed :installed, :upgradable do
           app_id = MacAppStore::Helpers::App.app_id_for?(new_resource.app_name)
           raise(Exceptions::InvalidAppName, new_resource.app_name) unless app_id
+
+          cmd = if new_resource.use_rtun
+                  include_recipe 'reattach-to-user-namespace'
+                  "reattach-to-user-namespace mas install #{app_id}"
+                else
+                  "mas install #{app_id}"
+                end
           execute "Upgrade #{new_resource.app_name} with Mas" do
-            command "mas install #{app_id}"
+            command cmd
             user new_resource.system_user
           end
         end