+++ recipes/default.rb
@@ -34,6 +34,9 @@ mac_app_store_mas 'default' do
   unless node['mac_app_store']['mas']['system_user'].nil?
     system_user node['mac_app_store']['mas']['system_user']
   end
+  unless node['mac_app_store']['mas']['use_rtun'].nil?
+    use_rtun node['mac_app_store']['mas']['use_rtun']
+  end
   action %i(install sign_in)
 end
 
@@ -42,6 +45,9 @@ node['mac_app_store']['apps'].to_h.each do |k, v|
   mac_app_store_app k do
     unless node['mac_app_store']['mas']['system_user'].nil?
       system_user node['mac_app_store']['mas']['system_user']
+    end
+    unless node['mac_app_store']['mas']['use_rtun'].nil?
+      use_rtun node['mac_app_store']['mas']['use_rtun']
     end
   end
 end